requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
langchain-core
langchain
langchain-google-genai
langchain-community
